Move RaindexRegistry core logic from js_api into common (#2470)
The registry business logic (file-format parsing, HTTP/data-URI fetching,
format validation, order/deployment metadata extraction, order-builder
construction) lived in the WASM-only `js_api` crate, so no native crate
could reuse it.

Extract it into a new platform-agnostic `raindex_common::registry` module:

- `RaindexRegistry` core struct + `RaindexRegistryError` carry all the
  fetching/parsing/validation logic with no WASM dependencies. The native
  `get_raindex_client` helper and all native (`httpmock`/`tokio`) tests move
  here too.
- `RaindexRegistry::get_order_builder` returns the common
  `RaindexOrderBuilder` core type and no longer threads a `js_sys::Function`
  callback.

`js_api::registry` keeps `DotrainRegistry` as a thin `#[wasm_bindgen]` /
`#[wasm_export]` wrapper over the core: it delegates to the inner type,
attaches the optional `js_sys::Function` state-update callback to the
returned builder, converts errors to `WasmEncodedError`/`JsValue`, and keeps
the WASM-only `getRaindexYaml`/`getRaindexClient` bindings. The exported JS
class name `DotrainRegistry` is unchanged, so the npm package and its TS/Svelte
consumers keep working.
