+++ crates/common/src/lib.rs
@@ -12,6 +12,7 @@ pub mod parsed_meta;
 pub mod raindex_client;
 pub mod raindex_order_builder;
 pub mod rainlang;
+pub mod registry;
 pub mod remove_order;
 #[cfg(not(target_family = "wasm"))]
 pub mod replays;
